backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
supabase==2.22.2
flask-restx==1.3.0
PyJWT==2.10.1
requests==2.31.0
gunicorn==21.2.0
httpx==0.27.0
